New code style
I like this because I can look at the top of the file and see exactly what functions/variables I can use in the template. You could also use `angular.extend` so you don't have to type `modal.` everywhere.

Also, doing things this way encourages you to make sure that you hoist all variables yourself (and all functions are auto-hoisted anyway).

Also, using function declarations like this assign the `function.name` property to the name of the function which is useful in debugging (though I believe the latest versions of Chrome will infer the name based on the name of the variable you assign the expression to, which is mind blowing).